minor-blockchain-1-base 1.extra modify

+++ params/opcodes.go
@@ -1,6 +1,6 @@
 package params
 
 var (
-	GCallValueTransfer = 9000
+	GCallValueTransfer = uint64(9000)
 	GtxxShardCost      = GCallValueTransfer // x-shard tx deposit gas
 )
